+++ HRLeaveManagement.Application/Features/LeaveRequest/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using HRLeaveManagement.Application.Features.LeaveRequest.Requests.Commands;
+﻿using HRLeaveManagement.Application.Exceptions;
+using HRLeaveManagement.Application.Features.LeaveRequest.Requests.Commands;
 using HRLeaveManagement.Application.Persistance.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -19,9 +21,12 @@ namespace HRLeaveManagement.Application.Features.LeaveRequest.Handlers.Commands
         }
         public async Task Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
-            var leaveRequest=  await  _leaveRequestRepository.Get(request.Id);
+            var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+            
+            if (leaveRequest == null)
+                throw new NotFoundException(nameof(leaveRequest), request.Id);
 
-           await  _leaveRequestRepository.Delete(leaveRequest);
+            await _leaveRequestRepository.Delete(leaveRequest);
         }
     }
 }
